// apps/messages-service/src/generated/grpc/src/grpc/messages.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.7
//   protoc               v3.19.1
// source: src/grpc/messages.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import Long = require("long");
import type { Metadata } from "@grpc/grpc-js";
import { Observable } from "rxjs";
import { Timestamp } from "../../google/protobuf/timestamp";

export const protobufPackage = "poslah.messages";

/** Указываем, что используем 3-ю версию синтаксиса */

/** Запрос для GetMessageAuthInfo */
export interface GetMessageAuthInfoRequest {
  chatId: string;
  nn: Long;
}

/** Ответ от GetMessageAuthInfo, содержащий только необходимые для авторизации поля */
export interface GetMessageAuthInfoResponse {
  authorId: string;
  createdAt: Timestamp | undefined;
  deletedAt: Timestamp | undefined;
}

export const POSLAH_MESSAGES_PACKAGE_NAME = "poslah.messages";

/** Сервис, который будет реализовывать наш messages-service */

export interface MessagesServiceClient {
  /** Метод для получения минимальной информации о сообщении для проверки прав доступа */

  getMessageAuthInfo(
    request: GetMessageAuthInfoRequest,
    metadata: Metadata,
    ...rest: any
  ): Observable<GetMessageAuthInfoResponse>;
}

/** Сервис, который будет реализовывать наш messages-service */

export interface MessagesServiceController {
  /** Метод для получения минимальной информации о сообщении для проверки прав доступа */

  getMessageAuthInfo(
    request: GetMessageAuthInfoRequest,
    metadata: Metadata,
    ...rest: any
  ): Promise<GetMessageAuthInfoResponse> | Observable<GetMessageAuthInfoResponse> | GetMessageAuthInfoResponse;
}

export function MessagesServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["getMessageAuthInfo"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("MessagesService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("MessagesService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const MESSAGES_SERVICE_NAME = "MessagesService";